Completely upgraded Rails to 3.1
I started out with an empty project and copied over only what I thought
were the relevant files, so I definitely missed some things. However,
the core app seems to work with a few bugs. Once those have been
identified and fixed, I can work on migrating the rest of the things
over.

I am new to Rails 3.1 and pretty new to Rails in general, but got
annoyed at being so far out of date that I had to update this open
source project. There is still a hell of a lot of optional work that
needs to be done to update Rails 2 practices to Rails 3 practices, and
I've already done the required ones. Here is a list of things to watch
out for:

- The h() function for stripping out HTML is now automatic, causing a
  lot of areas to have HTML that wouldn't render. This is solved by
  calling .html_safe on strings you know are safe.
- Some plugins were out of date and had to be removed. I found a few
  alternative gems that seem to function similarly. bbcode and search
  are the two areas that will behave differently, for good or for bad.
- Methods changed all over the place, leading to a lot of "<method name>
  rails 3" queries. I've found alternatives for most broken parts of the
  code, but you may stumble upon some "TODO"s.
- As mentioned before, I may have inadvertently deleted files that are
  not part of the core functionality. Nor have I tested these areas, so
  do not be suprised when things like the unit tests and Capistrano do
  not work.
- I made a few changes prior to updating to Rails 3, some of which may
  not be in the commit log. They almost exclusively have to do with
  avatars.

Please, don't hurt yourself and try to merge this commit, because it
will only end in pain. I would recommend rebasing and implementing all
of your changes one at a time. There is a lot more work to be done and
I'd like the help!